Integration of AES_PROCESS block into OLED program

+++ sdk/oled_temp/src/sd.c
@@ -115,7 +115,7 @@ bool create_test_bin(int choice)
 * @note     None
 *
 ******************************************************************************/
-bool write_to_file(const char *sdFile, const uint8_t *writeBuf, const uint32_t writeSize)
+bool write_to_file(const char *sdFile, const uint32_t *writeBuf, const uint32_t writeSize)
 {
 	FIL fil;
     UINT numOfBytesWritten; // Not used
@@ -155,7 +155,7 @@ bool write_to_file(const char *sdFile, const uint8_t *writeBuf, const uint32_t w
 * @note     None
 *
 ******************************************************************************/
-bool read_from_file(const char *sdFile, uint8_t *readBuf, uint32_t *readSize)
+bool read_from_file(const char *sdFile, uint32_t *readBuf, uint32_t *readSize)
 {
 	FIL fil;
     UINT numOfBytesRead; // Not used
